Minor refactoring and polishing
