+++ internal/proxy/proxy_server.go
@@ -122,7 +122,6 @@ func NewCentianProxy(globalConfig *config.GlobalConfig) (*CentianProxy, error) {
 // connections that pool currently owns.
 type UpstreamSession struct {
 	id              string
-	initialized     bool
 	upstreamServer  *mcp.Server                              // Server returned to the MCP SDK for this upstream session.
 	downstreamConns map[string]DownstreamConnectionInterface // Current downstream connections borrowed from the assigned pool.
 	registeredTools map[string]struct{}                      // Tools already registered on upstreamServer for this session.
@@ -295,28 +294,16 @@ func (p *MCPProxy) GetServerForRequest(r *http.Request) *mcp.Server {
 	return upstreamSession.upstreamServer
 }
 
-func getAuthHeaders(p *MCPProxy, r *http.Request) map[string]string {
-	authHeaders := make(map[string]string)
-	// Capture auth headers from upstream request for passthrough
-	// TODO: make these headers configurable
-	for _, h := range []string{"Authorization", "X-API-Key", "X-Auth-Token"} {
-		if p.server != nil && p.server.AuthHeader != "" && strings.EqualFold(h, p.server.AuthHeader) {
-			continue
-		}
-		if v := r.Header.Get(h); v != "" {
-			authHeaders[h] = v
-		}
-	}
-	return authHeaders
-}
-
 // createUpstreamSession captures the request state needed to attach a new
 // upstream MCP session to the correct reusable downstream pool.
 func (p *MCPProxy) createUpstreamSession(id string, r *http.Request, identityKey string) *UpstreamSession {
-	authHeaders := getAuthHeaders(p, r)
+	excludedAuthHeader := ""
+	if p.server != nil {
+		excludedAuthHeader = p.server.AuthHeader
+	}
+	authHeaders := getAuthHeaders(r.Header, excludedAuthHeader)
 	return &UpstreamSession{
 		id:              id,
-		initialized:     false,
 		downstreamConns: make(map[string]DownstreamConnectionInterface),
 		registeredTools: make(map[string]struct{}),
 		authHeaders:     authHeaders,
@@ -374,7 +361,6 @@ func (p *MCPProxy) initializeUpstreamSessionLocked(session *UpstreamSession) (*D
 
 	downstreamPool, reused := p.getOrCreateDownstreamPool(session)
 	session.downstreamConns = downstreamPool.downstreamConns
-	session.initialized = true
 	return downstreamPool, reused
 }
 
@@ -409,9 +395,8 @@ func (p *MCPProxy) ensureDownstreamConnectionsLocked(entry *DownstreamConnection
 			conn = p.newDownstreamConnection(serverName, connTemplate.config)
 			entry.downstreamConns[serverName] = conn
 		}
-		// TODO: why do we have 2 different fields for status here?
-		// 1. entry.connecting[serverName] -> true - see below
-		// 2. conn.GetStatus() == StatusConnecting
+		// entry.connecting tracks pool-level in-flight Connect goroutines so we
+		// do not start duplicate background connects for the same downstream.
 		if entry.connecting[serverName] {
 			continue
 		}
@@ -562,9 +547,9 @@ func (p *MCPProxy) connectDownstreamPoolConnection(
 	authHeaders map[string]string,
 ) {
 	serverName := conn.GetServerName()
-	// TODO: feels like we should set the status of the DownstreamConnection here somewhere...
+	// Connection status is owned by the DownstreamConnection implementation.
+	// The pool only tracks whether it currently has a background Connect call in flight.
 	if err := conn.Connect(context.Background(), authHeaders); err != nil {
-		// TODO: comment explaining this more in-depth
 		p.mu.Lock()
 		if entry, ok := p.downstreamPools[poolKey]; ok {
 			if current, exists := entry.downstreamConns[serverName]; exists && current == conn {
@@ -593,18 +578,20 @@ func (p *MCPProxy) connectDownstreamPoolConnection(
 		return
 	}
 
+	tools := conn.Tools()
+
 	p.toolRegMu.Lock()
 	for _, upstreamSession := range upstreamSessions {
 		if upstreamSession == nil || upstreamSession.upstreamServer == nil {
 			continue
 		}
-		for _, tool := range conn.Tools() {
+		for _, tool := range tools {
 			p.registerTool(upstreamSession, serverName, tool)
 		}
 	}
 	p.toolRegMu.Unlock()
 
-	common.LogInfo("MCPProxy[%s]: Connected pooled downstream %s with %d tools", p.name, sanitizeLogValue(serverName), len(conn.Tools()))
+	common.LogInfo("MCPProxy[%s]: Connected pooled downstream %s with %d tools", p.name, sanitizeLogValue(serverName), len(tools))
 }
 
 // NewMcpServer returns a new *mcp.Server based on the MCPProxy name.
@@ -857,21 +844,6 @@ func cloneAuthHeaders(headers map[string]string) map[string]string {
 	return cloned
 }
 
-func redactHeaders(headers http.Header) http.Header {
-	redacted := make(http.Header, len(headers))
-	for key, values := range headers {
-		if strings.EqualFold(key, "Authorization") || strings.EqualFold(key, "X-Centian-Auth") ||
-			strings.EqualFold(key, "X-API-Key") || strings.EqualFold(key, "X-Auth-Token") {
-			redacted[key] = []string{"<redacted>"}
-			continue
-		}
-		copied := make([]string, len(values))
-		copy(copied, values)
-		redacted[key] = copied
-	}
-	return redacted
-}
-
 func (p *MCPProxy) invalidatePooledDownstream(poolKey string) {
 	if poolKey == "" {
 		return
@@ -918,9 +890,7 @@ func RegisterEndpoint(endpoint string, proxy *MCPProxy, mux *http.ServeMux, opti
 		options,
 	)
 
-	var handler http.Handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		baseHandler.ServeHTTP(w, r)
-	})
+	var handler http.Handler = baseHandler
 	if proxy.server != nil && proxy.server.APIKeys != nil {
 		headerName := proxy.server.AuthHeader
 		if headerName == "" {
